Replaced stubForm store with Pinia
