--- gui/src/store/index.js
@@ -1,8 +1,0 @@
-import { createStore } from "vuex";
-import stubForm from "@/store/modules/stubForm";
-
-export default createStore({
-  modules: {
-    stubForm,
-  },
-});
